requirements.txt
requests~=2.32.5
pydantic~=2.12.5
langchain~=1.2.4
python-dotenv~=1.2.1
langchain-core~=1.2.7
langgraph~=1.0.6
PyYAML~=6.0.3
pydantic-settings~=2.12.0
SQLAlchemy~=2.0.45
fastapi~=0.128.0
neo4j~=6.1.0
sqlparse~=0.5.5
pandas~=2.3.3
numpy~=2.4.1
pymilvus~=2.6.6
psycopg2~=2.9.11
httpx~=0.28.1
deepagents~=0.3.6
pymongo~=4.16.0
bson~=0.5.10